Added alert for when Scribble is out of date.
Only moderators will see the alert, and it is below the moderator tools box, so even moderators may miss it, but it's better than nothing. This should help people know when to update Scribble. Don't worry, though, this will only be updated rarely, and will only be done for major update packages, not individual updates.